requirements.txt
accelerate==0.27.2
accept-paymob==1.0.0
annotated-types==0.6.0
anyio==4.2.0
arabic-reshaper==3.0.0
asgiref==3.7.2
async-timeout==4.0.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
diffusers==0.26.3
distro==1.9.0
Django==4.2.9
djangorestframework==3.14.0
exceptiongroup==1.2.0
fastapi==0.109.1
filelock==3.13.1
fsspec==2024.2.0
h11==0.14.0
hiredis==2.3.2
httpcore==1.0.3
httpx==0.26.0
huggingface-hub==0.21.3
idna==3.6
importlib-metadata==7.0.1
install==1.3.5
Jinja2==3.1.3
MarkupSafe==2.1.5
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
openai==1.13.3
packaging==23.2
pillow==10.2.0
psutil==5.9.8
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
python-bidi==0.4.2
python-ulid==1.1.0
pytz==2024.1
PyYAML==6.0.1
redis==4.6.0
redis-om==0.2.1
regex==2023.12.25
reportlab==4.1.0
requests==2.31.0
rtl==0.4.3
safetensors==0.4.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.35.1
stripe==8.5.0
sympy==1.12
tokenizers==0.15.2
tqdm==4.66.2
transformers==4.38.2
types-pyOpenSSL==24.0.0.20240130
types-redis==4.6.0.20240106
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.0.post1
zipp==3.17.0
